`local_uri` set to the Dispatcher.
Before setting the endpoint, the nodename (local_uri in the dtnd conf file) is set to the dispatcher as "nodeName".
The nodename is set before the endpoint eid so, if it is the case, the bundle already on the daemon do not disturb it. It is also better to fully set the details of the node before starting with the eid.